rework authentication
The session token seems to become invalid after 24h, and
redeploying the exporter with a manually updated config is
an unfeasible requirement for operation.

This adds chromedp to control a real browser for periodical
logins/session refreshes (see #1 for reasoning).

The refresh intervals are hard-coded for now, and are as
follows:

- at the very start, and then every 6 hours:
  perform an SSO login, and update the session cookie

- if the SSO login fails:
  reduce the interval to 30min (else reset to 6h)

- if the SSO login fails for 12h:
  abort the program